tabs-pane show non aliased names
